pyproject.toml
[tool.poetry]
name = "ai"
version = "0.1.0"
description = ""
authors = ["Juan Lee <ninejuan@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
transformers = "^4.46.2"
python-dotenv = "^1.0.1"
pymongo = "^4.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
